tender-hackathon/public_tender_collector/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "public-tender-collector"
version = "0.1.0"
description = "Compliance-first collector for publicly available Indian government tender documents."
authors = [{ name = "Procure Intelligence AP" }]
readme = "README.md"
license = { text = "Proprietary" }
# Spec asks for >=3.10,<3.13. We widen to <3.15 because the hackathon
# environment only has 3.13/3.14 installed; library compatibility on these
# minor versions has been verified manually for this prototype.
requires-python = ">=3.10,<3.15"
dependencies = [
  "httpx==0.27.2",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "pydantic==2.9.2",
  "pyyaml==6.0.2",
  "pandas==2.2.3",
  "sqlalchemy==2.0.36",
  "pymupdf==1.24.13",
  "python-docx==1.1.2",
  "openpyxl==3.1.5",
  "tenacity==8.5.0",
  "structlog==24.4.0",
  "typer==0.19.2",
  "rich==13.9.4",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-mock==3.14.0",
  "respx==0.21.1",
  "freezegun==1.5.1",
  "mypy==1.13.0",
  "ruff==0.7.4",
  "types-PyYAML==6.0.12.20240917",
  "pandas-stubs==2.2.3.241009",
]
playwright = [
  "playwright==1.45.0",
]

[project.scripts]
public-tender-collector = "main:app"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["collector*"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["data", "logs", "notebooks"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PL", "RUF"]
ignore = [
  "PLR0913",  # too many arguments — this is a config-heavy collector
  "PLR0911",  # too many returns — compliance.evaluate is intentionally enumerated
  "PLR0912",  # too many branches — same reason
  "PLR2004",  # magic numbers — thresholds are intentional
  "E501",     # long lines tolerated in comments / fixtures
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "B011", "PLR0913", "B904", "E741"]
# Typer's idiomatic API uses typer.Option(...) in argument defaults; B008
# would otherwise complain. PLR0915 (too many statements) is fine for the
# CLI orchestrator. B904 (raise from) is intentional for typer.Exit chains.
"main.py" = ["B008", "PLR0915", "B904", "E741"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_unreachable = true
disallow_untyped_decorators = false
plugins = ["pydantic.mypy"]
exclude = "(data|logs|notebooks|\\.venv)/"

[[tool.mypy.overrides]]
module = ["fitz", "openpyxl.*", "docx", "docx.*", "respx", "freezegun", "bs4", "bs4.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -ra"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::pydantic.PydanticDeprecatedSince20",
]
